Applied requested changes

+++ students/grzegorz_kowalik/lesson_01_basics/area_of_triangle.py
@@ -1,4 +1,4 @@
-a = int(input("Base: "))
-h = int(input("Height: "))
+a = float(input("Base: "))
+h = float(input("Height: "))
 
 print("Area: " + str(a*h/2))
